Cancel transcription tasks

+++ buzz/cache.py
@@ -28,4 +28,5 @@ class TasksCache:
             return []
 
     def clear(self):
-        os.remove(self.file_path)
+        if os.path.exists(self.file_path):
+            os.remove(self.file_path)
